[WIP] fix norm
