+++ includes/vsh.h
@@ -6,7 +6,7 @@
 /*   By: omulder <omulder@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/04/10 20:29:42 by jbrinksm       #+#    #+#                */
-/*   Updated: 2019/11/04 17:29:23 by jbrinksm      ########   odam.nl         */
+/*   Updated: 2019/11/05 11:03:50 by jbrinksm      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -838,11 +838,14 @@ int				jobs_mark_process_status(pid_t pid, int status);
 void			jobs_flush_process(t_proc *proc);
 void			jobs_launch_job(t_job *job);
 void			jobs_launch_proc(t_job *job, t_proc *proc,
-	int fds[3], int pipes[2]);
+					int fds[3], int pipes[2]);
 void			jobs_exec_builtin(t_proc *proc);
 int				jobs_exec_is_single_builtin_proc(t_proc *proc);
 
 void			jobs_force_job_state(t_job *job, t_proc_state state);
+void			jobs_launch_cleanup_after_proc(t_proc *proc, int fds[3],
+					int pipes[2]);
+void			jobs_launch_setup_stds(t_proc *proc, int fds[3], int pipes[2]);
 
 /*
 **----------------------------------shell---------------------------------------
@@ -1061,7 +1064,7 @@ int				exec_list(t_ast *ast, t_vshdata *data);
 int				exec_and_or(t_ast *ast, t_vshdata *data);
 int				exec_pipe_sequence(t_ast *ast, t_vshdata *data);
 int				exec_command(t_ast *ast, t_vshdata *g_data);
-void			exec_cmd(/*char **args, */t_vshdata *data);
+void			exec_cmd(t_vshdata *data);
 bool			exec_builtin(char **args, t_proc *proc);
 void			exec_external(char **args, t_vshdata *data, t_proc *proc);
 int				exec_find_binary(char *filename, t_vshdata *data,
@@ -1073,6 +1076,8 @@ int				exec_create_files(t_ast *ast);
 void			exec_add_pid_to_pipeseqlist(t_vshdata *data, pid_t pid);
 int				exec_redirs(t_ast *redirs);
 int				exec_assigns(t_ast *ast, t_vshdata *data, int env_type);
+char			**exec_create_process_args(t_ast *ast);
+bool			exec_command_contains_only_assign(t_ast *ast);
 
 /*
 **-----------------------------------signals------------------------------------
