+++ srcs/exec/exec_command.c
@@ -6,100 +6,15 @@
 /*   By: rkuijper <rkuijper@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/09/04 10:16:26 by rkuijper       #+#    #+#                */
-/*   Updated: 2019/11/04 13:44:53 by jbrinksm      ########   odam.nl         */
+/*   Updated: 2019/11/05 12:00:19 by jbrinksm      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <vsh.h>
 
-// static size_t	count_args(t_ast *ast)
-// {
-// 	t_ast	*probe;
-// 	size_t	i;
-
-// 	i = 0;
-// 	probe = ast;
-// 	while (probe != NULL && (probe->type == WORD || probe->type == ASSIGN))
-// 	{
-// 		i++;
-// 		probe = probe->left;
-// 	}
-// 	return (i);
-// }
-
-// static char		**malloc_args(int total_args)
-// {
-// 	char	**args;
-
-// 	if (total_args > 0)
-// 		args = (char**)ft_memalloc(sizeof(char*) * (total_args + 1));
-// 	else
-// 	{
-// 		args = (char**)ft_memalloc(sizeof(char*) * 2);
-// 		if (args == NULL)
-// 			return (NULL);
-// 		*args = ft_strnew(0);
-// 		if (*args == NULL)
-// 			ft_strarrdel(&args);
-// 	}
-// 	return (args);
-// }
-
-// static char		**create_args(t_ast *ast)
-// {
-// 	char	**args;
-// 	t_ast	*probe;
-// 	size_t	total_args;
-// 	size_t	i;
-
-// 	total_args = count_args(ast);
-// 	args = malloc_args(total_args);
-// 	if (args == NULL)
-// 		return (NULL);
-// 	i = 0;
-// 	probe = ast;
-// 	while (i < total_args)
-// 	{
-// 		args[i] = ft_strdup(probe->value);
-// 		if (args[i] == NULL)
-// 		{
-// 			ft_strarrdel(&args);
-// 			return (NULL);
-// 		}
-// 		probe = probe->left;
-// 		i++;
-// 	}
-// 	return (args);
-// }
-
-int				exec_command(t_ast *ast, t_vshdata *data)
+int		exec_command(t_ast *ast, t_vshdata *data)
 {
-	// char	**argv;
-
-	// if (expan_handle_variables(ast, data->envlst) == FUNCT_ERROR)
-	// 	return (FUNCT_ERROR);
-	// if (ast->type == WORD && expan_pathname(ast) == FUNCT_ERROR)
-	// 	return (FUNCT_ERROR);
-	// exec_quote_remove(ast);
-
-	// if (exec_command_contains_only_assign(ast) == true)
 	data->cur_node = ast;
-	// if (ast->type == WORD)
-	// 	data->current_redir_and_assign = ast->right;
-	// else
-	// 	data->current_redir_and_assign = ast;
-	// argv = create_args(ast);
-	// if (argv == NULL)
-		// return (FUNCT_ERROR);
-	exec_cmd(/*argv, */data);
-
-	// if (ast->type != WORD)
-	// {
-	// 	if (expan_handle_variables(ast, data->envlst) == FUNCT_ERROR)
-	// 		return (FUNCT_ERROR);
-	// 	exec_quote_remove(ast);
-	// 	if (exec_assigns(ast, data, ENV_LOCAL) == FUNCT_ERROR)
-	// 		return (FUNCT_ERROR);
-	// }
+	exec_cmd(data);
 	return (FUNCT_SUCCESS);
 }
