+++ srcs/jobs/jobs_launch.c
@@ -6,105 +6,13 @@
 /*   By: rkuijper <rkuijper@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/10/28 16:25:10 by rkuijper       #+#    #+#                */
-/*   Updated: 2019/11/04 17:54:30 by jbrinksm      ########   odam.nl         */
+/*   Updated: 2019/11/05 11:57:35 by jbrinksm      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vsh.h"
 #include "signal.h"
 #include <unistd.h>
-
-static size_t	count_args(t_ast *ast)
-{
-	t_ast	*probe;
-	size_t	i;
-
-	i = 0;
-	if (ast == NULL || ast->type != WORD)
-		return (0);
-	probe = ast;
-	while (probe != NULL && (probe->type == WORD || probe->type == ASSIGN))
-	{
-		i++;
-		probe = probe->left;
-	}
-	return (i);
-}
-
-static char		**malloc_args(int total_args)
-{
-	char	**args;
-
-	if (total_args > 0)
-		args = (char**)ft_memalloc(sizeof(char*) * (total_args + 1));
-	else
-	{
-		args = (char**)ft_memalloc(sizeof(char*) * 2);
-		if (args == NULL)
-			return (NULL);
-		*args = ft_strnew(0);
-		if (*args == NULL)
-			ft_strarrdel(&args);
-	}
-	return (args);
-}
-
-static char		**create_args(t_ast *ast)
-{
-	char	**args;
-	t_ast	*probe;
-	size_t	total_args;
-	size_t	i;
-
-	total_args = count_args(ast);
-	args = malloc_args(total_args);
-	if (args == NULL)
-		return (NULL);
-	i = 0;
-	probe = ast;
-	while (i < total_args)
-	{
-		args[i] = ft_strdup(probe->value);
-		if (args[i] == NULL)
-		{
-			ft_strarrdel(&args);
-			return (NULL);
-		}
-		probe = probe->left;
-		i++;
-	}
-	return (args);
-}
-
-static void	setup_stds(t_proc *proc, int fds[3], int pipes[2])
-{
-	if (proc->next != NULL)
-	{
-		if (pipe(pipes) < 0)
-		{
-			ft_eprintf("Could not pipe.\n");
-			exit(1);
-		}
-		fds[1] = pipes[1];
-	}
-	else
-		fds[1] = STDOUT_FILENO;
-}
-
-static void	cleanup_after_proc(t_proc *proc, int fds[3], int pipes[2])
-{
-	if (fds[0] != STDIN_FILENO && fds[0] != UNINIT)
-		close(fds[0]);
-	if (fds[1] != STDOUT_FILENO && fds[1] != UNINIT)
-		close(fds[1]);
-	fds[0] = pipes[0];
-	if (proc->binary != NULL)
-		ft_strdel(&proc->binary);
-	if (proc->argv != NULL)
-		ft_strarrdel(&proc->argv);
-	if (proc->env)
-		ft_memdel((void**)&proc->env);
-}
 
 static void	setup_fork(t_job *job, t_proc *proc, int fds[3], int pipes[2])
 {
@@ -126,20 +34,72 @@ static void	setup_fork(t_job *job, t_proc *proc, int fds[3], int pipes[2])
 	}
 }
 
-static bool		exec_command_contains_only_assign(t_ast *ast)
+static int	prepare_arguments_proc(t_ast *node, t_proc *proc, char ***argv)
 {
-	t_ast	*probe;
-
-	if (ast->type == WORD)
-		return (false);
-	probe = ast;
-	while (probe != NULL)
+	if (expan_handle_variables(node, g_data->envlst) == FUNCT_ERROR)
 	{
-		if (tool_is_redirect_tk(probe->type) == true)
-			return (false);
-		probe = probe->left;
+		//ERROR
+		ft_eprintf("ERROR3!\n");
+		return (FUNCT_ERROR);
 	}
-	return (true);
+	if (node->type == WORD && expan_pathname(node) == FUNCT_ERROR)
+	{
+		//ERROR
+		ft_eprintf("ERROR4!\n");
+		return (FUNCT_ERROR);
+	}
+	exec_quote_remove(node);
+	if (node->type == WORD)
+		proc->redir_and_assign = node->right;
+	else
+		proc->redir_and_assign = node;
+	*argv = exec_create_process_args(node);
+	if (*argv == NULL)
+	{
+		//ERROR
+		ft_eprintf("ERROR1!\n");
+		return (FUNCT_ERROR);
+	}
+	return (FUNCT_SUCCESS);
+}
+
+static int	prepare_settings_proc(t_job *job, t_proc *proc, t_ast *node,
+				char **argv)
+{
+	if (exec_command_contains_only_assign(node) == true)
+		proc->no_fork = true;
+	if (proc->no_fork == true && exec_assigns(proc->redir_and_assign,
+		g_data, ENV_LOCAL) == FUNCT_ERROR)
+	{
+		//ERROR
+		ft_eprintf("ERROR2!\n");
+		return (FUNCT_ERROR);
+	}
+	jobs_update_job_command(job, argv);
+	proc->argv = argv;
+	if (ft_strequ(proc->argv[0], "") == true)
+		proc->no_cmd = true;
+	if (exec_builtin(argv, proc) == false)
+		exec_external(argv, g_data, proc);
+	return (FUNCT_SUCCESS);
+}
+
+static int	handle_nonforked_builtin(t_job *job, t_proc *proc)
+{
+	if (job->bg == false && jobs_exec_is_single_builtin_proc(job->processes))
+	{
+		jobs_exec_builtin(job->processes);
+		env_remove_tmp(g_data->envlst);
+		jobs_finished_job(job, true);
+		return (FUNCT_FAILURE);
+	}
+	if (proc->is_builtin == false && proc->binary == NULL)
+	{
+		jobs_flush_job(job);
+		g_state->exit_code = 1;
+		return (FUNCT_FAILURE);
+	}
+	return (FUNCT_SUCCESS);
 }
 
 static int	launch_forked_job(t_job *job, int fds[3], int pipes[2])
@@ -152,72 +112,19 @@ static int	launch_forked_job(t_job *job, int fds[3], int pipes[2])
 	proc = job->processes;
 	while (proc != NULL)
 	{
-
-		/* */
+		argv = NULL;
 		node = proc->node;
-		if (expan_handle_variables(node, g_data->envlst) == FUNCT_ERROR)
-		{
-			//ERROR
-			ft_eprintf("ERROR3!\n");
+		if (prepare_arguments_proc(node, proc, &argv) == FUNCT_ERROR)
 			return (FUNCT_ERROR);
-		}
-		if (node->type == WORD && expan_pathname(node) == FUNCT_ERROR)
-		{
-			//ERROR
-			ft_eprintf("ERROR4!\n");
+		if (prepare_settings_proc(job, proc, node, argv) == FUNCT_ERROR)
 			return (FUNCT_ERROR);
-		}
-		exec_quote_remove(node);
-		if (node->type == WORD)
-			proc->redir_and_assign = node->right;
-		else
-			proc->redir_and_assign = node;
-		argv = create_args(node);
-		if (argv == NULL)
-		{
-			//ERROR
-			ft_eprintf("ERROR1!\n");
+		if (handle_nonforked_builtin(job, proc) != FUNCT_SUCCESS)
 			return (FUNCT_ERROR);
-		}
-		if (exec_command_contains_only_assign(node) == true)
-			proc->no_fork = true;
-		if (proc->no_fork == true)
-		{
-			if (exec_assigns(proc->redir_and_assign, g_data, ENV_LOCAL)
-				== FUNCT_ERROR)
-			{
-				//ERROR
-				ft_eprintf("ERROR2!\n");
-				return (FUNCT_ERROR);
-			}
-		}
-		jobs_update_job_command(job, argv);
-		proc->argv = argv;
-		if (ft_strequ(proc->argv[0], "") == true)
-			proc->no_cmd = true;
-		if (exec_builtin(argv, proc) == false)
-			exec_external(argv, g_data, proc);
-		/* */
-
-		if (job->bg == false && jobs_exec_is_single_builtin_proc(job->processes))
-		{
-			jobs_exec_builtin(job->processes);
-			env_remove_tmp(g_data->envlst);
-			jobs_finished_job(job, true);
-			return (FUNCT_FAILURE);
-		}
-
-		if (proc->is_builtin == false && proc->binary == NULL)
-		{
-			jobs_flush_job(job);
-			g_state->exit_code = 1;
-			return (FUNCT_FAILURE);
-		}
-		setup_stds(proc, fds, pipes);
+		jobs_launch_setup_stds(proc, fds, pipes);
 		pid = fork();
 		proc->pid = pid;
 		setup_fork(job, proc, fds, pipes);
-		cleanup_after_proc(proc, fds, pipes);
+		jobs_launch_cleanup_after_proc(proc, fds, pipes);
 		proc = proc->next;
 	}
 	return (FUNCT_SUCCESS);
